added add function
The add function has been updated with validation to ensure that only objects with expected keys can be added to the pokemonList array. The typeof operator checks that the input is an object, and the Object.keys method is used to check that the object has exactly the keys we expect (name, height, and type). If the input is invalid, an error message is logged to the console.